历史天气前后端连调完成

+++ ms-domain/src/main/java/com/ecjtu/domain/entity/WeatherData.java
@@ -47,7 +47,7 @@ public class WeatherData {
     /**
      * 风速
      */
-    private Integer windSpeed;
+    private Float windSpeed;
     /**
      * 风向
      */
